made changes to IMS POJOs mapping based on new IMS script
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@143 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/middleware/manager/InventoryDataManagerImpl.java
@@ -522,7 +522,7 @@ public class InventoryDataManagerImpl implements InventoryDataManager
 				
 				row.setEntityIdOfLot((Integer) result[1]);
 				
-				row.setActualLotBalance(((BigDecimal) result[2]).longValue());
+				row.setActualLotBalance(((Double) result[2]).longValue());
 				
 				Location location = germplasmManager.getLocationByID((Integer) result[3]);
 				row.setLocationOfLot(location);
